+++ kuStudy/kuStudy/extension/UIColor.swift
@@ -9,7 +9,7 @@
 import UIKit
 
 extension UIColor {
-    static var themeColor: UIColor {
+    static var theme: UIColor {
         return UIColor(hue: 359/359, saturation: 0.78, brightness: 0.83, alpha: 1)
     }
 }