+++ Release Notes.md
@@ -10,6 +10,7 @@
 
 ### Code improvements
 
+* App is now written in Swift 3.
 * Library names are localized.
 * Added shortened version of library names.